Rename module `util-common-base` -> `util-core`.
A little more succinct of a name, and it avoids confusion with the
similarly-named class `CommonBase`.
