+++ local-modules/util-common/JsonUtil.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { UtilityClass } from 'util-common-base';
+import { UtilityClass } from 'util-core';
 
 /**
  * JSON helper utilities.
