+++ local-modules/util-common/Random.js
@@ -7,7 +7,7 @@
 import crypto from 'crypto';
 
 import { TInt } from 'typecheck';
-import { UtilityClass } from 'util-common-base';
+import { UtilityClass } from 'util-core';
 
 /**
  * Character set used for ID strings. This is intended to be the set of 32 most
