+++ local-modules/util-core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "util-common-base",
+  "name": "util-core",
   "version": "1.0.0",
   "main": "main.js"
 }
